refactor(cli): expose auth via AuthClient and tighten api layout
- Colocate NewAuthClient with AuthClient in internal/api/auth.go and drop the standalone factory file so the auth surface stays in one place.

- Keep cmd depending on api.AuthClient while production still uses *Client under NewAuthClient, so tests can inject fakes without CRATER_HTTP_SIM.

- Rely on CRATER_HTTP_SIM only on the req transport in client.go for manual or integration-style failure simulation.

- Replace deprecated resp.IsSuccess() with resp.IsSuccessState() for req v3.

- Align SPEC / COMMANDS with internal/api file roles.
