+++ cli/internal/api/client.go
@@ -1,7 +1,11 @@
 package api
 
 import (
-	"fmt"
+	"context"
+	"io"
+	"net/http"
+	"os"
+	"strings"
 
 	"github.com/imroc/req/v3"
 )
@@ -13,53 +17,59 @@ type Response[T any] struct {
 	Message string `json:"msg"`
 }
 
-// LoginReq 登录请求体
-type LoginReq struct {
-	AuthMethod string `json:"auth"`
-	Username   string `json:"username"`
-	Password   string `json:"password"`
-}
-
-// LoginResp 登录响应数据
-type LoginResp struct {
-	AccessToken  string         `json:"accessToken"`
-	RefreshToken string         `json:"refreshToken"`
-	Context      AccountContext `json:"context"`
-	User         UserAttribute  `json:"user"`
-}
-
-// AccountContext 用户当前激活的上下文
-type AccountContext struct {
-	Queue        string `json:"queue"`
-	RoleQueue    int    `json:"roleQueue"`
-	RolePlatform int    `json:"rolePlatform"`
-	AccessQueue  int    `json:"accessQueue"`
-	AccessPublic int    `json:"accessPublic"`
-	Space        string `json:"space"`
-}
-
-// UserAttribute 用户属性
-type UserAttribute struct {
-	ID       int     `json:"id"`
-	Name     string  `json:"name"`
-	Nickname string  `json:"nickname"`
-	Email    *string `json:"email"`
-	Phone    *string `json:"phone"`
-	Teacher  *string `json:"teacher"`
-	Group    *string `json:"group"`
-}
-
-// Client Crater API 客户端
+// Client Crater API 客户端（真实 HTTP）
 type Client struct {
 	httpClient *req.Client
 	BaseURL    string
 }
 
+// EnvHTTPSim 为启用传输层错误模拟时读取的环境变量名；取值与行为见 docs/SPEC.md。
+const EnvHTTPSim = "CRATER_HTTP_SIM"
+
+// applyHTTPSim 按 CRATER_HTTP_SIM 在 req Transport 上注册拦截（仅影响经 NewClient 创建的客户端）。
+func applyHTTPSim(rc *req.Client) {
+	switch strings.TrimSpace(os.Getenv(EnvHTTPSim)) {
+	case "error404", "404":
+		wrapSim404(rc)
+	case "timeout", "hang":
+		wrapSimTimeout(rc)
+	default:
+	}
+}
+
+func wrapSim404(rc *req.Client) {
+	rc.GetTransport().WrapRoundTripFunc(func(_ http.RoundTripper) req.HttpRoundTripFunc {
+		return func(r *http.Request) (*http.Response, error) {
+			body := `{"code":404,"data":null,"msg":"simulated"}`
+			return &http.Response{
+				StatusCode:    http.StatusNotFound,
+				ProtoMajor:    1,
+				ProtoMinor:    1,
+				Status:        "404 Not Found",
+				Body:          io.NopCloser(strings.NewReader(body)),
+				Header:        http.Header{"Content-Type": []string{"application/json"}},
+				ContentLength: int64(len(body)),
+				Request:       r,
+			}, nil
+		}
+	})
+}
+
+func wrapSimTimeout(rc *req.Client) {
+	rc.GetTransport().WrapRoundTripFunc(func(_ http.RoundTripper) req.HttpRoundTripFunc {
+		return func(r *http.Request) (*http.Response, error) {
+			_ = r
+			return nil, context.DeadlineExceeded
+		}
+	})
+}
+
 // NewClient 初始化 API 客户端
 func NewClient(baseURL string) *Client {
-	c := req.C().SetBaseURL(baseURL)
+	rc := req.C().SetBaseURL(baseURL)
+	applyHTTPSim(rc)
 	return &Client{
-		httpClient: c,
+		httpClient: rc,
 		BaseURL:    baseURL,
 	}
 }
@@ -68,36 +78,4 @@ func NewClient(baseURL string) *Client {
 func (c *Client) SetToken(token string) *Client {
 	c.httpClient.SetCommonBearerAuthToken(token)
 	return c
-}
-
-// Login 用户登录
-func (c *Client) Login(username, password, mode string) (*LoginResp, error) {
-	var result Response[LoginResp]
-
-	resp, err := c.httpClient.R().
-		SetBody(&LoginReq{
-			AuthMethod: mode,
-			Username:   username,
-			Password:   password,
-		}).
-		SetSuccessResult(&result).
-		SetErrorResult(&result).
-		Post("/api/auth/login")
-
-	if err != nil {
-		return nil, fmt.Errorf("network error: %w", err)
-	}
-
-	if !resp.IsSuccess() {
-		if result.Message != "" {
-			return nil, fmt.Errorf("API error (%d): %s", result.Code, result.Message)
-		}
-		return nil, fmt.Errorf("API error: status code %d", resp.GetStatusCode())
-	}
-
-	if result.Code != 0 {
-		return nil, fmt.Errorf("API error (%d): %s", result.Code, result.Message)
-	}
-
-	return &result.Data, nil
 }
